+++ cmd/resumic/validate/doc.go
@@ -1,0 +1,1 @@
+package validate
